// src/app/shared/model/cbas-slik/req-add-trx-src-data-for-cbas-slik-obj.model.ts
export class ReqAddTrxSrcDataForCbasSlik{ 
    Name: string = '';
    CustType: number = 1;
    BirthDate: string = '';
    BirthPlace: string = '';
    KtpNo: string = '';
    AktaNo: string = '';
    Npwp: string = '';
    Gender: string = '';
    Address: string = '';
    MotherName: string = '';
    PurposeCode: string = '';

}